Added new error to the deleteUser function

+++ controllers/users.js
@@ -50,17 +50,21 @@ module.exports = {
   deleteUser: catchAsync(async (req, res, next) => {
     const { id } = req.params
     try {
-      await User.destroy({
-        where: {
-          id: id
-        }
-      })
-      const response = await User.findByPk(id, { paranoid: false })
-      endpointResponse({
-        res,
-        message: 'User deleted successfully',
-        body: response,
-      })
+      const response = await User.findByPk(id)
+      if (response) {
+        await User.destroy({
+          where: {
+            id: id
+          }
+        })
+        endpointResponse({
+          res,
+          message: 'User deleted successfully',
+          body: response,
+        })
+      }else {
+        throw new ErrorObject("Can't find the user you're looking for", 400);
+      }
     } catch (error) {
       const httpError = createHttpError(
         error.statusCode,
